requirements.txt
pdfplumber==0.10.3
sentence-transformers==2.7.0
transformers==4.35.2
huggingface_hub==0.19.4
faiss-cpu==1.7.4
streamlit==1.28.1
ollama==0.1.7
numpy==1.24.3
tiktoken==0.5.1
python-dotenv==1.0.0
torch==2.1.0
spacy==3.7.2
nltk==3.8.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
